refactor: move bookshelves into furniture package
